enhance: merge search_email to list_inbox
