+++ google/gmail/listAttachments.py
@@ -2,13 +2,13 @@ import os
 import json
 from googleapiclient.errors import HttpError
 
-from helpers import client, fetch_email_or_draft
-
+from apis.helpers import client
+from apis.messages import fetch_email_or_draft
 
 def main():
     email_id = os.getenv('EMAIL_ID')
-    if email_id is None and email_subject is None:
-        raise ValueError("Either email_id or email_subject must be set")
+    if email_id is None:
+        raise ValueError("Either email_id must be set")
 
     service = client('gmail', 'v1')
     try:
