+++ google/gmail/sendEmail.py
@@ -2,8 +2,8 @@ import os
 import asyncio
 from googleapiclient.errors import HttpError
 
-from helpers import client, create_message
-
+from apis.helpers import client
+from apis.messages import create_message
 
 async def main():
     to_emails = os.getenv('TO_EMAILS')
