+++ google/gmail/updateDraft.py
@@ -2,8 +2,8 @@ import os
 
 from googleapiclient.errors import HttpError
 
-from helpers import client, create_message
-
+from apis.helpers import client
+from apis.drafts import update_draft
 
 async def main():
     to_emails = os.getenv('TO_EMAILS')
@@ -38,18 +38,7 @@ async def main():
         print(err)
 
 
-async def update_draft(service, draft_id, to, cc, bcc, subject, body, attachments, reply_to_email_id=None, reply_all=False):
-    try:
-        message = await create_message(service=service, to=to, cc=cc, bcc=bcc, subject=subject, message_text=body, attachments=attachments, reply_to_email_id=reply_to_email_id, reply_all=reply_all)
-        updated_draft = {
-            'id': draft_id,
-            'message': message,
-        }
 
-        draft_response = service.users().drafts().update(userId='me', id=draft_id, body=updated_draft).execute()
-        print(f"Draft Id: {draft_response['id']} - Draft updated successfully!")
-    except HttpError as error:
-        print(f'An error occurred: {error}')
 
 
 if __name__ == "__main__":
